+++ openexchange-test/src/com/openexchange/ajax/folder/AbstractFolderCheckLimitTest.java
@@ -114,7 +114,7 @@ public class AbstractFolderCheckLimitTest extends AbstractEnhancedApiClientSessi
     private String getPrivateInfostoreFolder() throws Exception {
         ConfigApi configApi = new ConfigApi(quotaApiClient);
         ConfigResponse configNode = configApi.getConfigNode(Tree.PrivateInfostoreFolder.getPath());
-        return ((Integer) configNode.getData()).toString();
+        return (configNode.getData()).toString();
     }
 
     @Override
